Seperate auth, series and timer actions into seperate files.
