+++ src/app/components/login/login-form.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { authenticate } from "@/app/lib/actions";
+import { authenticate } from "@/app/lib/actions/actions";
 import { useFormState, useFormStatus } from "react-dom";
 
 const LoginForm = () => {
