--- src/app/lib/actions.ts
@@ -1,313 +1,0 @@
-"use server";
-
-import { auth, signIn } from "@/auth";
-import { Prisma } from "@prisma/client";
-import * as bcrypt from "bcryptjs";
-import { AuthError } from "next-auth";
-import { revalidatePath } from "next/cache";
-import { redirect } from "next/navigation";
-import { z } from "zod";
-import prisma from "./db";
-
-const SeriesFormSchema = z.object({
-  id: z.number(),
-  name: z.string(),
-  colour: z.string(),
-});
-
-const EditSeries = SeriesFormSchema.omit({ id: true });
-
-const TimerFormSchema = z.object({
-  id: z.number(),
-  name: z.string(),
-  colour: z.string(),
-  repeat: z.number(),
-  interval: z.number(),
-  main: z.number(),
-});
-
-const EditTimer = TimerFormSchema.omit({ id: true });
-
-// This is temporary until @types/react-dom is updated
-export type State = {
-  errors?: {
-    name?: string[];
-    colour?: string[];
-  };
-  message?: string | null;
-};
-
-export async function authenticate(_: string | undefined, formData: FormData) {
-  try {
-    await signIn("credentials", formData);
-  } catch (error) {
-    if (error instanceof AuthError) {
-      switch (error.type) {
-        case "CredentialsSignin":
-          return "Invalid credentials.";
-        default:
-          return "Something went wrong.";
-      }
-    }
-    throw error;
-  }
-}
-
-export async function register(_: string | undefined, formData: FormData) {
-  const email = formData.get("email") as string;
-  const password = formData.get("password") as string;
-  const confirmPassword = formData.get("confirm-password") as string;
-  if (password !== confirmPassword) {
-    return "Passwords do not match.";
-  }
-  try {
-    await prisma.user.create({
-      data: {
-        email,
-        password: await bcrypt.hash(password, 10),
-      },
-    });
-  } catch (error) {
-    if (error instanceof Prisma.PrismaClientKnownRequestError) {
-      if (error.code === "P2002") {
-        console.error(
-          "There is a unique constraint violation, a new user cannot be created with this email"
-        );
-      }
-    }
-    return "This email has already been registered.";
-  }
-  try {
-    await signIn("credentials", formData);
-  } catch (error) {
-    if (error instanceof AuthError) {
-      switch (error.type) {
-        case "CredentialsSignin":
-          return "Invalid credentials.";
-        default:
-          return "Something went wrong.";
-      }
-    }
-    throw error;
-  }
-  redirect("/series");
-}
-
-export async function editSeries(id: number, _: State, formData: FormData) {
-  let formColour = formData.get("colour") as string;
-  formColour = formColour.charAt(0).toLowerCase() + formColour.slice(1);
-  const validatedFields = EditSeries.safeParse({
-    name: formData.get("name"),
-    colour: formColour,
-  });
-  // If form validation fails, return errors early. Otherwise, continue.
-  if (!validatedFields.success) {
-    return {
-      errors: validatedFields.error.flatten().fieldErrors,
-      message: "Missing Fields. Failed to Edit Series.",
-    };
-  }
-  // Prepare data for insertion into the database
-  const { name, colour } = validatedFields.data;
-  // Update series.
-  try {
-    await prisma.series.update({
-      where: { id },
-      data: { name, colour },
-    });
-  } catch (error) {
-    console.error(error);
-    return {
-      message: "Database Error: Failed to Edit Series.",
-    };
-  }
-  // Revalidate the cache for the series page.
-  revalidatePath(`/series/${id}`);
-  return { message: "Edited Series." };
-}
-
-export async function deleteSeries(id: number) {
-  try {
-    await prisma.series.delete({
-      where: { id },
-    });
-  } catch (error) {
-    console.error(error);
-    return {
-      message: "Database Error: Failed to Delete Series.",
-    };
-  }
-  redirect("/series");
-  return { message: "Deleted Series." };
-}
-
-export async function createSeries(_: State, formData: FormData) {
-  // Validate user is logged in.
-  const session = await auth();
-  const userId = Number(session?.user?.id);
-  if (!userId) {
-    return {
-      message: "You must be logged in to create a series.",
-    };
-  }
-  // Process form data.
-  let formColour = formData.get("colour") as string;
-  formColour = formColour.charAt(0).toLowerCase() + formColour.slice(1);
-  const validatedFields = EditSeries.safeParse({
-    name: formData.get("name"),
-    colour: formColour,
-  });
-  // If form validation fails, return errors early. Otherwise, continue.
-  if (!validatedFields.success) {
-    return {
-      errors: validatedFields.error.flatten().fieldErrors,
-      message: "Missing Fields. Failed to Create Series.",
-    };
-  }
-  // Prepare data for insertion into the database
-  const { name, colour } = validatedFields.data;
-  // Create series.
-  try {
-    await prisma.series.create({
-      data: { name, colour, ownerId: userId },
-    });
-  } catch (error) {
-    console.error(error);
-    return {
-      message: "Database Error: Failed to Create Series.",
-    };
-  }
-  // Revalidate the cache for the series page.
-  revalidatePath("/series");
-  redirect("/series");
-  return { message: "Created Series." };
-}
-
-export async function editTimer(id: number, _: State, formData: FormData) {
-  let formColour = formData.get("colour") as string;
-  formColour = formColour.charAt(0).toLowerCase() + formColour.slice(1);
-  const validatedFields = EditTimer.safeParse({
-    name: formData.get("name"),
-    colour: formColour,
-    repeat: Number(formData.get("repeat")),
-    interval: Number(formData.get("interval")),
-    main: Number(formData.get("main")),
-  });
-  // If form validation fails, return errors early. Otherwise, continue.
-  if (!validatedFields.success) {
-    return {
-      errors: validatedFields.error.flatten().fieldErrors,
-      message: "Missing Fields. Failed to Edit Timer.",
-    };
-  }
-  // Prepare data for insertion into the database
-  const { name, colour, repeat, interval, main } = validatedFields.data;
-  // Update timer.
-  try {
-    await prisma.timer.update({
-      where: { id },
-      data: { name, colour, repeat, interval, main },
-    });
-  } catch (error) {
-    console.error(error);
-    return {
-      message: "Database Error: Failed to Edit Timer.",
-    };
-  }
-  // Revalidate the cache for the selected series page.
-  revalidatePath(`/series/${id}`);
-  return { message: "Created Timer." };
-}
-
-export async function deleteTimer(id: number, seriesId: number) {
-  prisma.$transaction(async (tx) => {
-    // Delete timer.
-    try {
-      await tx.timer.delete({
-        where: { id },
-      });
-    } catch (error) {
-      console.error(error);
-      return {
-        message: "Database Error: Failed to Delete Timer.",
-      };
-    }
-    // Update remaining timers positions.
-    try {
-      const timers = await tx.timer.findMany({ where: { seriesId } });
-      console.log("timers", timers.length, " for series ", seriesId);
-      timers.forEach(async (timer, index) => {
-        await prisma.timer.update({
-          where: { id: timer.id },
-          data: { position: index },
-        });
-      });
-    } catch (error) {
-      console.error(error);
-      return {
-        message: "Database Error: Failed to update timer position.",
-      };
-    }
-  });
-
-  revalidatePath(`/series/${seriesId}`);
-  return { message: "Deleted Timer." };
-}
-
-export async function createTimer(
-  id: number,
-  lastPosition: number,
-  _: State,
-  formData: FormData
-) {
-  // Validate user is logged in.
-  const session = await auth();
-  const userId = Number(session?.user?.id);
-  if (!userId) {
-    return {
-      message: "You must be logged in to create a timer.",
-    };
-  }
-  // Process form data.
-  let formColour = formData.get("colour") as string;
-  formColour = formColour.charAt(0).toLowerCase() + formColour.slice(1);
-  const validatedFields = EditTimer.safeParse({
-    name: formData.get("name"),
-    colour: formColour,
-    repeat: Number(formData.get("repeat")),
-    interval: Number(formData.get("interval")),
-    main: Number(formData.get("main")),
-  });
-  // If form validation fails, return errors early. Otherwise, continue.
-  if (!validatedFields.success) {
-    return {
-      errors: validatedFields.error.flatten().fieldErrors,
-      message: "Missing Fields. Failed to Create Series.",
-    };
-  }
-  // Prepare data for insertion into the database
-  const { name, colour, repeat, interval, main } = validatedFields.data;
-  // Create series.
-  try {
-    await prisma.timer.create({
-      data: {
-        name,
-        colour,
-        repeat,
-        interval,
-        main,
-        position: lastPosition,
-        seriesId: id,
-      },
-    });
-  } catch (error) {
-    console.error(error);
-    return {
-      message: "Database Error: Failed to Create Timer.",
-    };
-  }
-  // Revalidate the cache for the series page.
-  revalidatePath(`/series/${id}`);
-  redirect(`/series/${id}`);
-  return { message: "Created Timer." };
-}
