+++ src/app/lib/actions/types.ts
@@ -1,0 +1,8 @@
+// This is temporary until @types/react-dom is updated
+export type State = {
+  errors?: {
+    name?: string[];
+    colour?: string[];
+  };
+  message?: string | null;
+};
